requirements.txt
aiohttp==3.7.4.post0
anyio==3.5.0
apischema==0.16.6
async-timeout==3.0.1
attrs==21.4.0
base58==2.1.1
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.10
construct==2.10.67
construct-typing==0.5.2
discord.py==1.7.3
h11==0.12.0
httpcore==0.13.7
httpx==0.18.2
idna==3.3
jsonrpcclient==4.0.2
jsonrpcserver==5.0.6
jsonschema==3.2.0
multidict==6.0.2
OSlash==0.6.3
pycparser==2.21
PyNaCl==1.5.0
pyrsistent==0.18.1
python-dotenv==0.19.2
requests==2.27.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
solana==0.21.0
types-cachetools==4.2.9
typing-extensions==3.10.0.2
urllib3==1.26.8
websockets==10.1
yarl==1.7.2
